requirements.txt
# Pod Cleaner Dependencies
kubernetes>=28.1.0
python-dotenv>=1.0.0
schedule>=1.2.0
pyyaml>=6.0.1

# Logging and monitoring
structlog>=23.1.0
colorama>=0.4.6
prometheus_client>=0.17.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

